Added so users can add their email to mailing list. Working on checking the lists for the ones that subscribed to mailing added message for people who is trying to visit registered member only urls

+++ app/database/migrations/2014_03_10_002352_vBanUser.php
@@ -13,11 +13,11 @@ class VBanUser extends Migration {
 	public function up()
 	{
     Schema::create('vBanUser', function($table) {
-      $table->bigIncrements('id');
+      $table->increments('id');
       $table->bigInteger('community_id')->unique();
       $table->boolean('private_profile');
       $table->text('display_name');
-      $table->integer('steam_creation');
+      $table->bigInteger('steam_creation');
       $table->text('steam_avatar_url_big');
       $table->text('steam_avatar_url_small');
       $table->integer('vac_banned');
